+++ common/src/main/java/io/github/mortuusars/exposure/network/packet/server/CameraAddFrameC2SP.java
@@ -65,7 +65,6 @@ public record CameraAddFrameC2SP(InteractionHand hand, CompoundTag frame, List<U
         Camera.getCamera(player).ifPresentOrElse(camera -> {
             CameraItem cameraItem = camera.get().getItem();
             cameraItem.addFrame(serverPlayer, camera.get().getStack(), frame, getEntities(serverPlayer.serverLevel()));
-            cameraItem.onFrameAdded(serverPlayer, camera.get().getStack(), frame, getEntities(serverPlayer.serverLevel()));
         }, () -> Exposure.LOGGER.error("Cannot add frame. Player is not using a Camera."));
 
         ItemStack cameraStack = player.getItemInHand(hand);